backend/requirements.txt
# NOTE: These pins are validated on Python 3.11.
# If you're on Python 3.13, create a 3.11 environment first.
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.3
gymnasium==0.29.1
stable-baselines3==2.2.1
pygame==2.6.1
numpy==1.26.3
Pillow==10.2.0
openai==1.10.0
